bmad-backend/requirements.txt
langchain-openai
langchain
langgraph
python-dotenv
fastapi
uvicorn[standard]
pydantic
langchain-core
figmapy
requests
pyjwt
markdown
beautifulsoup4
lxml
langchain_community